requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
numpy==2.1.3
openai==1.54.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
